.XLSX File Example w/ Code
VLOOKUP is a Google Sheets and Microsoft Excel function that retrieves information based on a set of parameters. The functionality can be explained like this (in pseudocode)... =VLOOKUP(PlantPollination Plant KeyA2, range from PlantReference KeyA2 to MethodC10, PlantReference number of columns in range, FALSE) . This pseudocode logic can be applied to across multiple sheets for an array of information (as shown with the inclusion of the LocationReference sheet).